// workspace/enums.ts
export enum Tasks_FieldIds {
  description_275 = "6773a73ebd84fc36f14d9275",
  assigned_to_user_276 = "6773a73ebd84fc36f14d9276",
  deadline_277 = "6773a73ebd84fc36f14d9277",
  time_spent_278 = "6773a73ebd84fc36f14d9278",
  company_d68 = "6786c139d9d0a2495e369d68",
  ohjausryhm_544 = "67b63ba53188de27bde4e544",
  admin_5a5 = "67b63bbd3188de27bde4e5a5",
  source_890 = "68cc55f8661806bb5d9fe890",
  assigned_to_5cc = "68d6240a63ea029c6cb955cc",
  priority_524 = "68da452ab0988db31d470524",
  unique_id_729 = "68da4585b0988db31d470729",
  metadata_89c = "68dd070af66abedb4d5b789c",
  additional_raw_info_920 = "68dd0717f66abedb4d5b7920",
  subtasks_e83 = "68dd84df30628eed73bdde83",
  subtasks_f47 = "68dd851830628eed73bddf47",
}

// Phase IDs enum for workflow: Tasks
// Use these constants instead of hardcoded IDs in phasesOrder

export enum Tasks_PhaseIds {
  to_do_279 = "6773a73ebd84fc36f14d9279",
  doing_27a = "6773a73ebd84fc36f14d927a",
  done_27b = "6773a73ebd84fc36f14d927b",
  accounting_fe7 = "68dbceb2c2a9a981f7576fe7",
  payroll_07d = "68dbcec2c2a9a981f757707d",
  financial_statement_47a = "68dbcfa7c2a9a981f757747a",
}
// ...
export enum Personnel_FieldIds {
  phone_number_a53 = "678e01095dd886b1d2208a53",
  email_a5d = "678e010b5dd886b1d2208a5d",
  yritys_cd0 = "678e01565dd886b1d2208cd0",
  user_48c = "68d6237963ea029c6cb9548c",
}
// ...
export enum Subtasks_FieldIds {
  main_task_d3a = "68dd849d30628eed73bddd3a",
  status_d7f = "68dd84b930628eed73bddd7f",
}
// ...
export enum WorkflowIds {
  companies_a51 = "677393544e9c3c0478e35a51",
  tasks_274 = "6773a73ebd84fc36f14d9274",
  time_tracker_ad3 = "6785882fec0d9f8f3f487ad3",
  personnel_815 = "678a1ce0c5ab52aefeaae815",
  materiaalitesitykset_11c = "6799de94f361d9d32544911c",
  ohjausryhmt_2d4 = "67b63ada3188de27bde4e2d4",
  reoccurring_tasks_81f = "68c004aa097b8b294325c81f",
  subtasks_cc4 = "68dd848730628eed73bddcc4",
  kyc_223 = "68e7650aa713093242dbf223",
  contacts_ffb = "68ef3391ec49f35fffc1affb",
  shareholders_13f = "68ef3408ec49f35fffc1b13f",
  shares_transactions_c4c = "68ef3595ec49f35fffc1bc4c",
  marketing_4c3 = "68ef3afe642706a2e322d4c3",
  kyc_signspace_form_6ad = "68f74c6bda248873fd8896ad",
  systems_519 = "690776634ee2003ee256f519",
  systems_in_use_609 = "6907769c4ee2003ee256f609",
  notes_rekna_monthly_1f4 = "69689a250af694ce3738d1f4",
}